pyproject.toml
[tool.poetry]
name = "ubot"
version = "1.0.0"
description = "A bot made for the Factorio Discord server."
authors = ["arielbeje <arielbeje@protonmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "main.py"},
    {include = "cogs"}
]

[tool.poetry.dependencies]
python = "^3.7"
ago = "^0.0.93"
aiohttp = ">=3.3.0,<3.5.0"
aiosqlite = "^0.9.0"
asyncpg = "^0.18.3"
beautifulsoup4 = "^4.7"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", branch = "rewrite" }
feedparser = "^5.2"
fuzzywuzzy = "^0.17.0"
python-Levenshtein = "^0.12.0"
pytz = "^2018.9"
tomd = "^0.1.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
rope = "^0.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
